requirements.txt
disnake
gTTS
loguru
python-dotenv
PyNaCl